decimal was troublesome, float is adequate

+++ html/model.py
@@ -1,12 +1,11 @@
 from pony.orm import db_session, Database, Required, Json
-from decimal import Decimal
 
 db = Database()
 
 class Bike(db.Entity):
     # tags = Required(Json)
-    lon = Required(Decimal)
-    lat = Required(Decimal)
-
+    lon = Required(float)
+    lat = Required(float)
+    
     def __repr__(self):
         return "<bike %s %s %s>" % (self.id, self.lon, self.lat)
